autorig-online/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
httpx==0.26.0
authlib==1.3.0
itsdangerous==2.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
slowapi==0.1.9
aiofiles==23.2.1
resend>=2.0.0

# Poster NSFW classification (NudeNet ONNX)
nudenet==3.4.2
opencv-python-headless==4.13.0.92
onnxruntime==1.24.4
numpy>=1.24.0

# YouTube Data API uploads
google-api-python-client==2.159.0
google-auth==2.35.0
google-auth-httplib2==0.2.0

